+++ server/src/services/sentences_service.ts
@@ -208,12 +208,18 @@ export const deleteAll = async (sentence: Sentence[]) => {
   return;
 };
 
-export const deleteAllYesterday = async (yesterday: dayjs.Dayjs) => {
-  try {
-    await model.UsersRecommendedSentences.destroy({
-      where: {
-        createdAt: { [Op.lt]: yesterday.format('YYYY-MM-DD HH:mm') },
-      },
-    });
-  } catch (err) {}
+export const readAllYesterday = async (yesterday: dayjs.Dayjs) => {
+  const yesterdaySentences: UsersRecommendedSentences[] = await model.UsersRecommendedSentences.findAll({
+    where: {
+      createdAt: { [Op.lt]: yesterday.format('YYYY-MM-DD HH:mm') },
+    },
+  });
+  return yesterdaySentences;
+};
+
+export const deleteAllYesterday = async (yesterdaySentences: UsersRecommendedSentences[]) => {
+  yesterdaySentences.forEach(async (element) => {
+    await element.destroy();
+  });
+  return;
 };
